rename get-sources-jenkins.sh scripts to get-sources.sh; reverse the -p flag :: dashboard

+++ codeready-workspaces-dashboard/get-sources.sh
@@ -5,13 +5,16 @@ verbose=1
 scratchFlag=""
 doRhpkgContainerBuild=1
 forceBuild=0
-forcePull=0
+# NOTE: pullAssets (-p) flag uses opposite behaviour to some other get-sources.sh scripts;
+# here we want to collect assets during sync-to-downsteam (using get-sources.sh -n -p)
+# so that rhpkg build is simply a brew wrapper (using get-sources.sh -f)
+pullAssets=0
 
 while [[ "$#" -gt 0 ]]; do
 	case $1 in
+		'-p'|'--pull-assets') pullAssets=1; shift 0;;
 		'-n'|'--nobuild') doRhpkgContainerBuild=0; shift 0;;
 		'-f'|'--force-build') forceBuild=1; shift 0;;
-		'-p'|'--force-pull') forcePull=1; shift 0;;
 		'-s'|'--scratch') scratchFlag="--scratch"; shift 0;;
 	esac
 	shift 1
@@ -30,30 +33,33 @@ function logn()
 	fi
 }
 
-# step one - build the builder image
-BUILDER=$(command -v podman || true)
-if [[ ! -x $BUILDER ]]; then
-	echo "[WARNING] podman is not installed, trying with docker"
-	BUILDER=$(command -v docker || true)
+if [[ ${pullAssets} -eq 1 ]]; then 
+	# step one - build the builder image
+	BUILDER=$(command -v podman || true)
 	if [[ ! -x $BUILDER ]]; then
-			echo "[ERROR] must install docker or podman. Abort!"; exit 1
+		echo "[WARNING] podman is not installed, trying with docker"
+		BUILDER=$(command -v docker || true)
+		if [[ ! -x $BUILDER ]]; then
+				echo "[ERROR] must install docker or podman. Abort!"; exit 1
+		fi
 	fi
+
+	BOOTSTRAPFILE=./build/dockerfiles/rhel.Dockerfile
+	# with yarn 2, no need to change the dockerfile (unlike with go vendoring or yarn 1)
+	tag=$(pwd);tag=${tag##*/}
+	${BUILDER} build . -f ${BOOTSTRAPFILE} --target builder -t ${tag}:bootstrap # --no-cache
+
+	# step two - extract cache folder to tarball (let's hope there's nothing arch-specific we need here!)
+	${BUILDER} run --rm --entrypoint sh ${tag}:bootstrap -c 'tar -pzcf - node_modules' > "asset-node-modules-cache.tgz"
+	${BUILDER} rmi ${tag}:bootstrap
+    outputFiles=asset-node-modules-cache.tgz
 fi
 
-BOOTSTRAPFILE=./build/dockerfiles/rhel.Dockerfile
-# with yarn 2, no need to change the dockerfile (unlike with go vendoring or yarn 1)
-tag=$(pwd);tag=${tag##*/}
-${BUILDER} build . -f ${BOOTSTRAPFILE} --target builder -t ${tag}:bootstrap # --no-cache
-
-# step two - extract cache folder to tarball (let's hope there's nothing arch-specific we need here!)
-${BUILDER} run --rm --entrypoint sh ${tag}:bootstrap -c 'tar -pzcf - node_modules' > "asset-node-modules-cache.tgz"
-${BUILDER} rmi ${tag}:bootstrap
-
-outputFiles=asset-node-modules-cache.tgz
-if [[ ${outputFiles} ]]; then
-	log "[INFO] Upload new sources:${outputFiles}"
+if [[ $(git diff-index HEAD --) ]] || [[ ${pullAssets} -eq 1 ]]; then
+	git add sources Dockerfile .gitignore || true
+	log "[INFO] Upload new sources: ${outputFiles}"
 	rhpkg new-sources ${outputFiles}
-	log "[INFO] Commit new sources from:${outputFiles}"
+	log "[INFO] Commit new sources from: ${outputFiles}"
 	if [[ $(git commit -s -m "[get sources] ${COMMIT_MSG}" sources Dockerfile .gitignore) == *"nothing to commit, working tree clean"* ]]; then 
 		log "[INFO] No new sources, so nothing to build."
 	elif [[ ${doRhpkgContainerBuild} -eq 1 ]]; then
@@ -88,3 +94,4 @@ $ERRORS
 		log "[INFO] No new sources, so nothing to build."
 	fi
 fi
+
